+++ src/react-components/src/components/01-atoms/Text/Headings/RNSectionTitle.stories.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+
+import RNSectionTitle from "./RNSectionTitle";
+import { action } from "@storybook/addon-actions";
+
+export default {
+  title: "Research Now Section Title",
+  component: RNSectionTitle
+};
+
+export const researchNowSectionTitle = () => <RNSectionTitle titleId="hello-page-title" sectionUrl="hello" blockName="blockname" modifiers={["hi", "bye"]}/>;