+++ src/react-components/src/components/01-atoms/Text/Headings/RNSectionTitle.tsx
@@ -1,0 +1,17 @@
+/* AT-79 Page Title - Level 1 */
+import * as React from "react";
+import bem from "../../../../utils/bem";
+import Heading from "./Heading";
+
+export interface RNSectionTitleProps {
+  titleId: string;
+  sectionUrl: string;
+  modifiers?: string[];
+  blockName?: string;
+}
+
+export default function RNSectionTitle(props: RNSectionTitleProps) {
+    const { titleId, sectionUrl, modifiers, blockName } = props;
+    const baseClass = "rn-section-title";
+    return <a className={baseClass} href={sectionUrl}>Research<b>NOW</b></a>;
+}