+++ src/react-components/src/components/03-organisms/Headers/HeaderWithSearch/HeaderWithSearch.tsx
@@ -3,18 +3,23 @@ import * as React from "react";
 import SearchBar from "../../../02-molecules/SearchBar/SearchBar";
 import UnderlineLink from "../../../01-atoms/Links/UnderlineLink";
 import bem from "../../../../utils/bem";
+import BodyText from "../../../01-atoms/Text/Text/BodyText";
+import SectionTitle from "../../../02-molecules/Headings/SectionTitle";
+import RNSectionTitle from "../../../01-atoms/Text/Headings/RNSectionTitle";
 
 
 export interface WithSearchProps {
   modifiers?: [];
   blockName?: string;
+  titleId: string;
   searchBarId: string;
   searchButtonId: string;
   dropdownId: string;
+  headingText: string;
   hasError?: boolean;
   errorMessage?: string;
   searchDropdownOptions: string[];
-  advancedSearchMessage: JSX.Element;
+  advancedSearchLink: string;
   selectChangeHandler: (event: React.FormEvent) => void;
   selectBlurHandler: (event: React.FormEvent) => void;
   searchSubmitHandler: (event: React.MouseEvent) => void;
@@ -23,16 +28,19 @@ export interface WithSearchProps {
 
 export default function HeaderWithSearch(props: React.PropsWithChildren<WithSearchProps>) {
 
-  const { blockName, searchBarId, searchButtonId, hasError, errorMessage,
-    dropdownId, searchDropdownOptions, advancedSearchMessage,
-    selectChangeHandler, selectBlurHandler, searchSubmitHandler, textChangeHandler } = props; const baseClass = "header-with-image-right";
+  const { blockName, titleId, searchBarId, searchButtonId, hasError, errorMessage,
+    dropdownId, headingText, searchDropdownOptions, advancedSearchLink,
+    selectChangeHandler, selectBlurHandler, searchSubmitHandler, textChangeHandler } = props;
+    const base_class = "search-header";
 
   return (
-    <div className={bem(baseClass)}>
+    <div className={bem(base_class)}>
+      <RNSectionTitle titleId={titleId} sectionUrl="#sectionUrl" blockName={base_class} />
       <SearchBar
         searchBarId={searchBarId}
         buttonId={searchButtonId}
         dropdownId={dropdownId}
+        textFieldAriaLabel="blah"
         dropdownOptions={searchDropdownOptions}
         hasError={hasError}
         errorMessage={errorMessage}
@@ -40,7 +48,9 @@ export default function HeaderWithSearch(props: React.PropsWithChildren<WithSear
         selectChangeHandler={selectChangeHandler}
         searchSubmitHandler={searchSubmitHandler}
         searchChangeHandler={textChangeHandler} />
-      <UnderlineLink>{advancedSearchMessage}</UnderlineLink>
+        <div className={bem("promo-text", [], base_class)}>
+          <UnderlineLink url={advancedSearchLink} >Advanced Search</UnderlineLink>
+        </div>
     </div>
   );
 }