+++ src/react-components/CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease.  When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ========
+## PRERELEASE
+### Added 
+- `Header With Search` 
+- `RN Section Title`
+
+### Changed
+- `Link` missing URL error message
 
 ## [0.0.6] - 2019-12-31
 ### Added